backend/ITSbackend/requirements.txt
asgiref==3.3.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
Django==3.0.5
django-filter==2.4.0
django-storages==1.9.1
djangorestframework==3.12.2
djongo==1.3.3
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
idna==2.10
importlib-metadata==3.3.0
Markdown==3.3.3
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.0
pymongo==3.11.2
python-resize-image==1.1.19
pytz==2020.5
requests==2.25.1
rsa==4.7
six==1.15.0
sqlparse==0.2.4
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
django-cors-headers
